pyproject.toml
[project]
name = "d4c"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.3.2",
    "python-dotenv>=1.0.1",
    "aiohttp>=3.9.0",
]

[tool.uv]
package = false

[dependency-groups]
dev = [
    "ruff>=0.15.6",
]
